requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.9
Django==4.2.3
djangorestframework==3.14.0
fastapi==0.99.1
ffmpy==0.3.0
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gpt-index==0.4.24
gradio==3.35.2
gradio_client==0.2.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.2
idna==3.4
Jinja2==3.1.2
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
kiwisolver==1.4.4
langchain==0.0.118
langchainplus-sdk==0.0.20
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pycryptodome==3.18.0
pydantic==1.10.11
pydub==0.25.1
Pygments==2.15.1
pyparsing==3.0.9
PyPDF2==3.0.1
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rpds-py==0.7.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
sqlparse==0.4.4
starlette==0.27.0
tenacity==8.2.2
tiktoken==0.4.0
toolz==0.12.0
tqdm==4.65.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
uc-micro-py==1.0.2
urllib3==2.0.3
uvicorn==0.22.0
websockets==11.0.3
yarl==1.9.2
